test: pull data into separate file
