--- src/test/query/highlights.graphql.ts
@@ -1,20 +1,0 @@
-import { gql } from 'apollo-server-core';
-
-export const GET_HIGHLIGHTS = gql`
-  query GetHighlights($itemId: ID) {
-    _entities(representations: { id: $itemId, __typename: "SavedItem" }) {
-      ... on SavedItem {
-        annotations {
-          highlights {
-            id
-            patch
-            version
-            quote
-            _createdAt
-            _updatedAt
-          }
-        }
-      }
-    }
-  }
-`;
